refactor: move ProofData and LeafData out of claim_note.rs into dedicated proof_data module
